backend/requirements.txt
Deprecated==1.2.14
GitPython==3.1.42
Jinja2==3.1.3
MarkupSafe==2.1.5
PyMySQL==1.0.2
PyPika==0.48.9
PyYAML==6.0
Pygments==2.17.2
SQLAlchemy==1.4.44
aiohttp==3.8.3
aiomysql==0.1.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.6.2
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
asynctest==0.13.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boilerpy3==1.0.7
cachetools==5.3.3
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
chroma-haystack==0.15.0
chroma-hnswlib==0.7.3
chromadb==0.4.19
click==8.1.3
coloredlogs==15.0.1
distlib==0.3.6
distro==1.9.0
dnspython==2.2.1
elastic-transport==8.4.0
elasticsearch==8.5.2
email-validator==1.3.0
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-utils==0.2.1
feedparser==6.0.11
filelock==3.8.0
flatbuffers==24.3.7
frozenlist==1.3.3
fsspec==2024.3.0
gitdb==4.0.11
google-auth==2.28.2
googleapis-common-protos==1.63.0
greenlet==2.0.1
grpcio==1.62.1
h11==0.14.0
haystack-ai==2.0.0
haystack-bm25==1.0.2
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
identify==2.5.9
idna==3.4
importlib-metadata==6.11.0
importlib_resources==6.3.1
joblib==1.3.2
jsonschema==3.0.2
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
lazy-imports==0.3.1
llmsherpa==0.1.4
lxml==5.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
motor==3.1.1
mpmath==1.3.0
multidict==6.0.2
networkx==3.2.1
newspaper4k==0.9.3
nltk==3.8.1
nodeenv==1.7.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.14.1
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
overrides==7.7.0
packaging==21.3
pandas==2.2.1
pendulum==3.0.0
pgvector==0.2.5
pillow==10.2.0
pip==24.0
platformdirs==2.5.4
posthog==3.5.0
pre-commit==2.20.0
protobuf==4.25.3
psycopg==3.1.18
psycopg-binary==3.1.18
pulsar-client==3.4.0
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.2
pydantic_core==2.16.3
pydeck==0.8.1b0
pymongo==4.3.3
pyparsing==3.0.9
pypdf==4.1.0
pyrsistent==0.19.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-multipart==0.0.9
pytz==2024.1
redis==4.3.5
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
setuptools==69.2.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
time-machine==2.14.0
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.5.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.19.0
virtualenv==20.17.0
watchdog==4.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.8.1
zipp==3.10.0